feat: Extract task computing into a strategy interface (#13690)
